+++ src/main/java/de/maxhenkel/camera/integration/jei/JEIPlugin.java
@@ -1,9 +1,14 @@
 package de.maxhenkel.camera.integration.jei;
 
 import de.maxhenkel.camera.Main;
+import de.maxhenkel.camera.RecipeImageCloning;
 import de.maxhenkel.camera.gui.*;
 import mezz.jei.api.IModPlugin;
+import mezz.jei.api.recipe.category.extensions.IExtendableRecipeCategory;
+import mezz.jei.api.recipe.category.extensions.vanilla.crafting.ICraftingCategoryExtension;
 import mezz.jei.api.registration.IGuiHandlerRegistration;
+import mezz.jei.api.registration.IVanillaCategoryExtensionRegistration;
+import net.minecraft.item.crafting.ICraftingRecipe;
 import net.minecraft.util.ResourceLocation;
 
 @mezz.jei.api.JeiPlugin
@@ -21,6 +26,12 @@ public class JEIPlugin implements IModPlugin {
         registration.addGuiScreenHandler(LecternAlbumScreen.class, NoJEIGuiProperties::new);
         registration.addGuiScreenHandler(ResizeFrameScreen.class, NoJEIGuiProperties::new);
         registration.addGuiScreenHandler(CameraScreen.class, NoJEIGuiProperties::new);
+    }
+
+    @Override
+    public void registerVanillaCategoryExtensions(IVanillaCategoryExtensionRegistration registration) {
+        IExtendableRecipeCategory<ICraftingRecipe, ICraftingCategoryExtension> craftingCategory = registration.getCraftingCategory();
+        craftingCategory.addCategoryExtension(RecipeImageCloning.class, ImageCopyExtension::new);
     }
 
 }